Fixed Unicode build of configuration utility, and fixed a compiler warning in MidiSynth.cpp.

+++ driver/src/MidiSynth.cpp
@@ -137,7 +137,7 @@ public:
 		}
 		numBytes = (bitDepth / 8) * numChannels;
 
-		PCMWAVEFORMAT wFormat = {WAVE_FORMAT_PCM, (WORD)numChannels, sampleRate, sampleRate * numBytes, (WORD)numBytes, bitDepth};
+		PCMWAVEFORMAT wFormat = {WAVE_FORMAT_PCM, (WORD)numChannels, sampleRate, sampleRate * numBytes, (WORD)numBytes, (WORD)bitDepth};
 
 		// Open waveout device
 		int wResult = waveOutOpen(&hWaveOut, WAVE_MAPPER, (LPWAVEFORMATEX)&wFormat, callback, (DWORD_PTR)&midiSynth, callbackType);
